fix: vendored solidity casts

+++ test/vendored/balancer/vault/PoolRegistry.sol
@@ -120,7 +120,7 @@ abstract contract PoolRegistry is ReentrancyGuard, VaultAuthorization {
 
         serialized |= bytes32(uint256(nonce));
         serialized |= bytes32(uint256(specialization)) << (10 * 8);
-        serialized |= bytes32(uint256(pool)) << (12 * 8);
+        serialized |= bytes32(uint256(uint160(pool))) << (12 * 8);
 
         return serialized;
     }
@@ -133,7 +133,7 @@ abstract contract PoolRegistry is ReentrancyGuard, VaultAuthorization {
     function _getPoolAddress(bytes32 poolId) internal pure returns (address) {
         // 12 byte logical shift left to remove the nonce and specialization setting. We don't need to mask,
         // since the logical shift already sets the upper bits to zero.
-        return address(uint256(poolId) >> (12 * 8));
+        return address(uint160(uint256(poolId) >> (12 * 8)));
     }
 
     /**
