+++ test/vendored/balancer/vault/VaultAuthorization.sol
@@ -74,7 +74,7 @@ abstract contract VaultAuthorization is
 
     constructor(IAuthorizer authorizer)
         // The Vault is a singleton, so it simply uses its own address to disambiguate action identifiers.
-        Authentication(bytes32(uint256(address(this))))
+        Authentication(bytes32(uint256(uint160(address(this)))))
         SignaturesValidator("Balancer V2 Vault")
     {
         _setAuthorizer(authorizer);
